added some scene object tests
